requirements.txt
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
gspread==5.11.0
oauth2client==4.1.3
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
scikit-learn==1.3.2
statsmodels==0.14.1
openpyxl==3.1.2
cachetools==5.3.2
cryptography==42.0.5
Jinja2==3.1.3
scipy==1.11.4